docker: and new Fedora 30 image description and remove 28
From the Fedora 30 image onwards we will split off some of special use
parts. This will keep the base image smaller and allow us to use it in
the FROM directive in teh more specific dockerfile descriptions. Namely
this will be mingw deps and coverity tools as a start. Later on we would
also have a testing image with things like exactness data.